allrequirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-bootstrap3==1.2.1
aldryn-categories==1.0.4
aldryn-common==1.0.4
aldryn-newsblog==1.3.2
aldryn-people==1.2.2
aldryn-reversion==1.1.0
aldryn-style==1.1.4
aldryn-translation-tools==0.2.1
appdirs==1.4.3
Babel==2.4.0
backport-collections==0.1
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.8.18
django-appconf==1.0.2
django-appdata==0.1.6
django-classy-tags==0.8.0
django-cms==3.4.2
django-durationfield==0.5.4
django-filer==1.2.7
django-formtools==2.0
django-ipware==1.1.6
django-mptt==0.8.7
django-parler==1.7
django-phonenumber-field==1.3.0
django-polymorphic==1.0.2
django-reversion==1.10.2
django-secretballot==0.6.0
django-sekizai==0.10.0
django-sortedm2m==1.3.3
django-taggit==0.22.0
django-treebeard==4.1.0
django-vote==2.1.4
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.2.0
djangocms-column==1.7.0
djangocms-forms==0.2.5
djangocms-googlemap==1.0.2
djangocms-installer==0.9.6
djangocms-link==2.1.1
djangocms-picture==2.0.4
djangocms-revolutionslider==0.9
djangocms-slider==0.2.4
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
easy-thumbnails==2.3
hashids==1.2.0
html5lib==0.9999999
jsonfield==2.0.1
lxml==3.7.3
olefile==0.44
packaging==16.8
phonenumbers==8.4.1
phonenumberslite==8.4.1
Pillow==4.1.0
polls==0.0.4
pyparsing==2.2.0
python-dateutil==2.6.0
python-slugify==1.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
sorl-thumbnail==12.3
tablib==0.11.4
tzlocal==1.3.2
Unidecode==0.4.20
YURL==0.13
